set_stacklimit_restart: print rlim_t with a 64-bit conversion, not %zu
stack_limits.rlim_cur, .rlim_max and RLIM_INFINITY are all rlim_t, which is
64-bit on 32-bit Linux while size_t - and so %zu - is 32-bit there. Printing
them with %zu is a wrong-size varargs read: the values come out as garbage, and
every argument after the first mismatched one is taken from the wrong stack
offset.

On 64-bit builds rlim_t and size_t are both 64-bit, so nothing looks wrong
there, which is why this has gone unnoticed. The function is reachable in
ordinary use, called from Mlucas.c:4062.

Cast to uint64 and use PRIu64, as the rest of the file does for 64-bit
quantities.

Verified on i686 under Debian 13: -Wformat warnings go from 5 to 0.

+++ src/util.c
@@ -1894,8 +1894,8 @@ void set_stacklimit_restart(char *argv[])
 		fprintf(stderr, "Call to getrlimit() failed.\n");
 		ASSERT(0, "Exiting.");
 	}
-	printf("Old stack_limits: cur = %zu, max = %zu, [RLIM_INFINITY = %zu]\n",
-	       stack_limits.rlim_cur, stack_limits.rlim_max, RLIM_INFINITY);
+	printf("Old stack_limits: cur = %" PRIu64 ", max = %" PRIu64 ", [RLIM_INFINITY = %" PRIu64 "]\n",
+	       (uint64)stack_limits.rlim_cur, (uint64)stack_limits.rlim_max, (uint64)RLIM_INFINITY);
 
 	if (stack_limits.rlim_cur == stack_limits.rlim_max)
 		return;
@@ -1905,8 +1905,8 @@ void set_stacklimit_restart(char *argv[])
 		fprintf(stderr, "Call to setrlimit() failed.\n");
 		ASSERT(0, "Exiting.");
 	}
-	printf("New stack_limits: cur = %zu, max = %zu\n",
-	       stack_limits.rlim_cur, stack_limits.rlim_max);
+	printf("New stack_limits: cur = %" PRIu64 ", max = %" PRIu64 "\n",
+	       (uint64)stack_limits.rlim_cur, (uint64)stack_limits.rlim_max);
 
 	if(execvp(argv[0], argv)) {
 		fprintf(stderr, "Call to execvp() failed.\n");
